+++ sickbeard/providers/blutopia.py
@@ -149,7 +149,7 @@ class BlutopiaProvider(generic.TorrentProvider):
                                     tr, {'seed': r'circle-up', 'leech': r'circle-down', 'size': r'fa-file'})
                                 seeders, leechers, size = [tryInt(n, n) for n in [
                                     cells[head[x]].get_text().strip() for x in 'seed', 'leech', 'size']]
-                                if self._peers_fail(mode, seeders, leechers):
+                                if self._reject_item(seeders, leechers):
                                     continue
 
                                 title = tr.find('a', href=rc['info'])