+++ sickbeard/providers/torrenting.py
@@ -93,7 +93,7 @@ class TorrentingProvider(generic.TorrentProvider):
                                 head = head if None is not head else self._header_row(tr)
                                 seeders, leechers, size = [tryInt(n, n) for n in [
                                     cells[head[x]].get_text().strip() for x in 'seed', 'leech', 'size']]
-                                if None is tr.find('a', href=rc['cats']) or self._peers_fail(mode, seeders, leechers):
+                                if None is tr.find('a', href=rc['cats']) or self._reject_item(seeders, leechers):
                                     continue
 
                                 info = tr.find('a', href=rc['info'])